requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv
google-genai
openpyxl>=3.1.0
